requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
pymongo[srv]==4.3.3
gunicorn==20.1.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
dnspython==2.3.0
flask-talisman==1.0.0
Pillow==9.5.0
certifi==2023.7.22
python-dotenv==1.0.0
